backend/requirements.txt
# requirements.txt

fastapi
uvicorn
scikit-learn
joblib
pandas
numpy
yfinance
requests
alpaca-py
pymongo
dnspython
python-dotenv
tensorflow
bcrypt
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx